Solve problem 14 part 1
The problem requires, for each hash, keeping the next thousand hashes in
memory to see if any of them contain a repeated character present in
that hash.

This lends itself to two easy optimizations:
1) We can only compute each hash once by keeping a ring buffer of hashes
2) We can keep an index of which hashes contain five-character sequences
to short circuit future comparisons after the first time.

It's likely I'll need more optimizations for part two of the problem,
but for now this seems to work well enough in under a minute.