Solve problem 11 part 2
This required an optimization to solve part 2 in my 4GB of RAM --
instead of storing each previous board state, just store an incrementing
integer for each move and the hash of the board in the `seen_states`
set. This should allow ruby to GC the old board states, with only a
minor false positive risk from multiple hashes colliding.

This took a few hours to run, so this was definitely the hardest problem
so far to efficiently solve. I'll have to come back and try to implement
A* properly.