Add 2022/15
      --------Part 1---------   --------Part 2---------
Day       Time    Rank  Score       Time    Rank  Score
 15   00:46:25    2982      0   01:50:23    3030      0

Pretty hard to get this one right. I missed the first one because of
bounds checking - I mistakenly limited the min/max x coordinates of row
2,000,000 to be based on the distance from the closest sensor, not the
max distance from any sensor, so any overlapping bigger exclusion zones
would not have counted.

Part 2 was even trickier. I tried figuring it out with a skipping
approach (i.e. when reaching a sensor's exclusion zone, use math to
figure out how much of it can be skipped, and then skip to the end of
it). But I couldn't get the math to line up. I found this approach on
Reddit and it was easier to implement and just worked.